+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/StatefulProcessorHandleImpl.scala
@@ -21,10 +21,6 @@ import java.util.UUID
 
 import scala.collection.mutable
 
-import org.json4s.{DefaultFormats, Formats, JValue}
-import org.json4s.JsonAST.{JBool, JString}
-import org.json4s.JsonDSL._
-
 import org.apache.spark.TaskContext
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.Encoder
@@ -70,52 +66,6 @@ class QueryInfoImpl(
 
   override def toString: String = {
     s"QueryInfo(queryId=$queryId, runId=$runId, batchId=$batchId)"
-  }
-}
-
-sealed trait StateVariableType
-
-case object ValueState extends StateVariableType
-case object ListState extends StateVariableType
-case object MapState extends StateVariableType
-
-object StateVariableType {
-  def withName(name: String): StateVariableType = name match {
-    case "ValueState" => ValueState
-    case "ListState" => ListState
-    case "MapState" => MapState
-    case _ => throw new IllegalArgumentException(s"Unknown state type: $name")
-  }
-}
-class StateVariableInfo(
-    val stateName: String,
-    val stateType: StateVariableType,
-    val isTtlEnabled: Boolean
-) {
-  override def toString: String = {
-    s"StateVariableInfo(stateName=$stateName, stateType=$stateType, isTtlEnabled=$isTtlEnabled)"
-  }
-
-  def jsonValue: JValue = {
-    ("stateName" -> JString(stateName)) ~
-      ("stateType" -> JString(stateType.toString)) ~
-      ("isTtlEnabled" -> JBool(isTtlEnabled))
-
-  }
-}
-
-object StateVariableInfo {
-  implicit val formats: Formats = DefaultFormats
-  def fromJson(json: Any): List[StateVariableInfo] = {
-    val stateVariables = json.asInstanceOf[List[Map[String, Any]]]
-    // Extract each JValue to StateVariableInfo
-    stateVariables.map { stateVariable =>
-      new StateVariableInfo(
-        stateVariable("stateName").asInstanceOf[String],
-        StateVariableType.withName(stateVariable("stateType").asInstanceOf[String]),
-        stateVariable("isTtlEnabled").asInstanceOf[Boolean]
-      )
-    }
   }
 }
